scanmypdf-api/requirements.txt
anyio==3.5.0
asgiref==3.5.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
fastapi==0.73.0
ghostscript==0.7
google-api-core==2.4.0
google-auth==2.6.0
google-cloud-core==2.2.2
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
h11==0.13.0
httptools==0.3.0
idna==3.3
importlib-metadata==4.10.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.1
uvloop==0.16.0
Wand==0.6.7
watchgod==0.7
websockets==10.1
zipp==3.7.0
